Chunk 1 - Added SecurityCoverageInformation component to relationship view
