+++ opencti-platform/opencti-front/lang/front/de.json
@@ -779,6 +779,7 @@
   "Coverage name": "Name der Abdeckung",
   "Coverage recurrence (every x)": "Wiederholung der Deckung (alle x)",
   "Coverage score (0-100)": "Deckungsgrad (0-100)",
+  "Coverage scores": "Erfassungsbereich Scores",
   "Coverage update periodicity": "Periodizität der Aktualisierung des Versicherungsschutzes",
   "Coverage validity period": "Gültigkeitsdauer des Versicherungsschutzes",
   "Covered entity": "Abgedeckte Entität",
@@ -2411,6 +2412,7 @@
   "IPv4 addresses": "IPv4-Adressen",
   "IPv6 address": "IPv6-Adresse",
   "IPv6 addresses": "IPv6-Adressen",
+  "Is covered": "Ist abgedeckt",
   "is empty": "ist leer",
   "Is family": "Ist Familie",
   "Is family?": "Ist Familie?",
