+++ opencti-platform/opencti-front/lang/front/en.json
@@ -779,6 +779,7 @@
   "Coverage name": "Coverage name",
   "Coverage recurrence (every x)": "Coverage recurrence (every x)",
   "Coverage score (0-100)": "Coverage score (0-100)",
+  "Coverage scores": "Coverage scores",
   "Coverage update periodicity": "Coverage update periodicity",
   "Coverage validity period": "Coverage validity period",
   "Covered entity": "Covered entity",
@@ -2411,6 +2412,7 @@
   "IPv4 addresses": "IPv4 addresses",
   "IPv6 address": "IPv6 address",
   "IPv6 addresses": "IPv6 addresses",
+  "Is covered": "Is covered",
   "is empty": "is empty",
   "Is family": "Is family",
   "Is family?": "Is family?",
