+++ opencti-platform/opencti-front/lang/front/es.json
@@ -779,6 +779,7 @@
   "Coverage name": "Nombre de la cobertura",
   "Coverage recurrence (every x)": "Recurrencia de cobertura (cada x)",
   "Coverage score (0-100)": "Puntuación de la cobertura (0-100)",
+  "Coverage scores": "Puntuaciones de cobertura",
   "Coverage update periodicity": "Periodicidad de actualización de la cobertura",
   "Coverage validity period": "Periodo de validez de la cobertura",
   "Covered entity": "Entidad cubierta",
@@ -2411,6 +2412,7 @@
   "IPv4 addresses": "Direcciones IPv4",
   "IPv6 address": "Dirección IPv6",
   "IPv6 addresses": "Direcciones IPv6",
+  "Is covered": "Está cubierto",
   "is empty": "está vacío",
   "Is family": "Es una familia",
   "Is family?": "¿Es una familia?",
