+++ opencti-platform/opencti-front/lang/front/fr.json
@@ -779,6 +779,7 @@
   "Coverage name": "Nom de la couverture",
   "Coverage recurrence (every x)": "Récurrence de la couverture (tous les x)",
   "Coverage score (0-100)": "Score de couverture (0-100)",
+  "Coverage scores": "Scores de couverture",
   "Coverage update periodicity": "Périodicité de la mise à jour de la couverture",
   "Coverage validity period": "Période de validité de la couverture",
   "Covered entity": "Entité couverte",
@@ -2411,6 +2412,7 @@
   "IPv4 addresses": "Adresses IPv4",
   "IPv6 address": "Adresse IPv6",
   "IPv6 addresses": "Adresses IPv6",
+  "Is covered": "Est couvert",
   "is empty": "est vide",
   "Is family": "Est une famille",
   "Is family?": "Est une famille ?",
