+++ opencti-platform/opencti-front/lang/front/it.json
@@ -779,6 +779,7 @@
   "Coverage name": "Nome della copertura",
   "Coverage recurrence (every x)": "Ricorrenza della copertura (ogni x)",
   "Coverage score (0-100)": "Punteggio della copertura (0-100)",
+  "Coverage scores": "Punteggi di copertura",
   "Coverage update periodicity": "Periodicità di aggiornamento della copertura",
   "Coverage validity period": "Periodo di validità della copertura",
   "Covered entity": "Entità coperta",
@@ -2411,6 +2412,7 @@
   "IPv4 addresses": "Indirizzi IPv4",
   "IPv6 address": "Indirizzo IPv6",
   "IPv6 addresses": "Indirizzi IPv6",
+  "Is covered": "È coperto",
   "is empty": "è vuoto",
   "Is family": "È famiglia",
   "Is family?": "È famiglia?",
