+++ opencti-platform/opencti-front/lang/front/ja.json
@@ -780,6 +780,7 @@
   "Coverage name": "カバレッジ名",
   "Coverage recurrence (every x)": "カバレッジの再発（x回ごと）",
   "Coverage score (0-100)": "カバレッジ・スコア（0～100）",
+  "Coverage scores": "カバレッジ・スコア",
   "Coverage update periodicity": "カバレッジの更新周期",
   "Coverage validity period": "補償有効期間",
   "Covered entity": "対象エンティティ",
@@ -2413,6 +2414,7 @@
   "IPv4 addresses": "IPv4アドレス",
   "IPv6 address": "IPv6アドレス",
   "IPv6 addresses": "IPv6アドレス",
+  "Is covered": "カバーされている",
   "is empty": "空です",
   "Is family": "ファミリ",
   "Is family?": "ファミリかどうか",
