+++ opencti-platform/opencti-front/lang/front/ko.json
@@ -780,6 +780,7 @@
   "Coverage name": "커버리지 이름",
   "Coverage recurrence (every x)": "적용 범위 반복(매 x마다)",
   "Coverage score (0-100)": "커버리지 점수 (0-100)",
+  "Coverage scores": "커버리지 점수",
   "Coverage update periodicity": "적용 범위 업데이트 주기",
   "Coverage validity period": "커버리지 유효 기간",
   "Covered entity": "적용 대상 엔터티",
@@ -2412,6 +2413,7 @@
   "IPv4 addresses": "IPv4 주소",
   "IPv6 address": "IPv6 주소",
   "IPv6 addresses": "IPv6 주소",
+  "Is covered": "적용됨",
   "is empty": "비어 있음",
   "Is family": "가족 여부",
   "Is family?": "가족입니까?",
