+++ opencti-platform/opencti-front/lang/front/ru.json
@@ -780,6 +780,7 @@
   "Coverage name": "Название покрытия",
   "Coverage recurrence (every x)": "Повторяемость покрытия (каждый x)",
   "Coverage score (0-100)": "Оценка покрытия (0-100)",
+  "Coverage scores": "Баллы за покрытие",
   "Coverage update periodicity": "Периодичность обновления покрытия",
   "Coverage validity period": "Срок действия покрытия",
   "Covered entity": "Охваченная сущность",
@@ -2413,6 +2414,7 @@
   "IPv4 addresses": "IPv4-адреса",
   "IPv6 address": "IPv6-адрес",
   "IPv6 addresses": "Адреса IPv6",
+  "Is covered": "Охвачен ли",
   "is empty": "пустой",
   "Is family": "Семья",
   "Is family?": "Семья?",
