+++ opencti-platform/opencti-front/lang/front/zh.json
@@ -780,6 +780,7 @@
   "Coverage name": "覆盖范围名称",
   "Coverage recurrence (every x)": "覆盖重复性（每 x 次）",
   "Coverage score (0-100)": "覆盖范围得分（0-100）",
+  "Coverage scores": "覆盖分数",
   "Coverage update periodicity": "保险更新周期",
   "Coverage validity period": "保险有效期",
   "Covered entity": "覆盖实体",
@@ -2413,6 +2414,7 @@
   "IPv4 addresses": "IPv4地址",
   "IPv6 address": "IPv6地址",
   "IPv6 addresses": "IPv6地址",
+  "Is covered": "是否承保",
   "is empty": "为空",
   "Is family": "是同一家族",
   "Is family?": "是否是同一家族?",
