+++ opencti-platform/opencti-front/src/private/components/analyses/security_coverages/SecurityCoverageInformation.tsx
@@ -1,208 +1,41 @@
-import React, { FunctionComponent } from 'react';
-import makeStyles from '@mui/styles/makeStyles';
-import { useTheme } from '@mui/styles';
-import { Avatar, Tooltip } from '@mui/material';
-import { BugReportOutlined, HourglassEmpty, Person, ShieldOutlined, TrackChangesOutlined } from '@mui/icons-material';
-import Chart from '@components/common/charts/Chart';
-import { ApexOptions } from 'apexcharts';
+import Label from '@common/label/Label';
+import { Box } from '@mui/material';
+import SecurityCoverageScores from './SecurityCoverageScores';
 import { useFormatter } from '../../../../components/i18n';
-import { isEmptyField, isNotEmptyField } from '../../../../utils/utils';
-import { donutChartOptions } from '../../../../utils/Charts';
-import type { Theme } from '../../../../components/Theme';
-import { capitalizeFirstLetter } from '../../../../utils/String';
-
-const useStyles = makeStyles((theme: Theme) => ({
-  charts: {
-    display: 'flex',
-    gap: theme.spacing(3),
-    flexWrap: 'wrap',
-  },
-  chartItem: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    gap: theme.spacing(1),
-  },
-  chart: {
-    position: 'absolute',
-    top: -5,
-    left: -5,
-  },
-  chartContainer: {
-    position: 'relative',
-    overflow: 'hidden',
-    width: 60,
-    height: 60,
-    padding: 4,
-  },
-  iconOverlay: {
-    fontSize: 24,
-    position: 'absolute',
-    top: 22,
-    left: 22,
-  },
-  scoreText: {
-    fontSize: 14,
-    fontWeight: 600,
-    color: theme.palette.text?.primary || '#ffffff',
-  },
-  coverageName: {
-    fontSize: 12,
-    color: theme.palette.text?.secondary || '#999999',
-    textAlign: 'center',
-  },
-}));
+import ItemBoolean from '../../../../components/ItemBoolean';
 
 interface SecurityCoverageInformationProps {
   coverage_information: ReadonlyArray<{
     readonly coverage_name: string;
     readonly coverage_score: number;
   }> | null | undefined;
-  variant?: 'header' | 'details' | 'matrix';
 }
 
-const SecurityCoverageInformation: FunctionComponent<SecurityCoverageInformationProps> = ({ coverage_information, variant = 'header' }) => {
+const SecurityCoverageInformation = ({ coverage_information }: SecurityCoverageInformationProps) => {
   const { t_i18n } = useFormatter();
-  const theme = useTheme<Theme>();
-  const classes = useStyles();
-  const genOpts = (score: number | null) => {
-    let chartColors = [theme.palette.action?.disabled ?? '#ffffff'];
-    let labels = [t_i18n('Unknown')];
-    let series = [score ?? 100];
-    if (isNotEmptyField(score)) {
-      chartColors = [theme.palette.success.main ?? '', theme.palette.error.main ?? ''];
-      labels = [t_i18n('Success'), t_i18n('Failure')];
-      series = [score, 100 - score];
-    }
-    const options = donutChartOptions(
-      theme,
-      labels,
-      'bottom',
-      false,
-      chartColors,
-      false,
-      false,
-      true,
-      false,
-      65,
-      false,
-    ) as ApexOptions;
-    return { series, options };
-  };
+  const hasCoverageInformation = (coverage_information ?? []).length > 0;
 
-  const ICON_MAP: Record<string, (size: number) => React.ReactElement> = {
-    DETECTION: (size) => <TrackChangesOutlined style={{ fontSize: size }} />,
-    PREVENTION: (size) => <ShieldOutlined style={{ fontSize: size }} />,
-    VULNERABILITY: (size) => <BugReportOutlined style={{ fontSize: size }} />,
-    MANUAL: (size) => <Person style={{ fontSize: size }} />,
-  };
-
-  const getIconFromName = (name: string, iconSize: number): React.ReactElement => {
-    const iconFactory = ICON_MAP[name.toUpperCase()];
-    return iconFactory
-      ? iconFactory(iconSize)
-      : <HourglassEmpty style={{ fontSize: iconSize }} />;
-  };
-
-  // Original variant for header or matrix (compact)
-  if (variant === 'header' || variant === 'matrix') {
-    const size = variant === 'matrix' ? 28 : 40;
-    const chartSize = variant === 'matrix' ? 38 : 50;
-    const iconSize = variant === 'matrix' ? 12 : 18;
-    const iconPosition = variant === 'matrix' ? 13 : 17;
-    if (isEmptyField(coverage_information)) {
-      const { options, series } = genOpts(null);
-      return (
-        <div className={classes.chartContainer} style={{ width: size, height: size }}>
-          <div className={classes.chart}>
-            <Chart options={options} series={series} type="donut" width={chartSize} height={chartSize} />
-            <Tooltip title="Pending" placement="bottom">
-              <Avatar className={classes.iconOverlay} sx={{ bgcolor: 'transparent', width: iconSize, height: iconSize }} style={{ top: iconPosition, left: iconPosition, fontSize: iconSize - 2 }}>
-                <span style={{ color: theme.palette.text?.primary }}><HourglassEmpty style={{ fontSize: iconSize }} /></span>
-              </Avatar>
-            </Tooltip>
-          </div>
-        </div>
-      );
-    }
-    return (
-      <div style={{ display: 'flex' }}>
-        {(coverage_information ?? []).map((coverageResult) => {
-          const { options, series } = genOpts(coverageResult.coverage_score);
-          return (
-            <div key={coverageResult.coverage_name} className={classes.chartContainer} style={{ width: size, height: size, padding: variant === 'matrix' ? 2 : 4 }}>
-              <div className={classes.chart}>
-                <Chart options={options} series={series} type="donut" width={chartSize} height={chartSize} />
-                <Tooltip title={`${coverageResult.coverage_name} ${coverageResult.coverage_score}/100`} placement="bottom">
-                  <Avatar className={classes.iconOverlay} sx={{ bgcolor: 'transparent', width: iconSize, height: iconSize }} style={{ top: iconPosition, left: iconPosition - 1, fontSize: iconSize - 2 }}>
-                    <span style={{ color: theme.palette.text?.primary }}>{getIconFromName(coverageResult.coverage_name, iconSize)}</span>
-                  </Avatar>
-                </Tooltip>
-              </div>
-            </div>
-          );
-        })}
-      </div>
-    );
-  }
-
-  // Details variant with scores
-  if (isEmptyField(coverage_information)) {
-    const { options, series } = genOpts(null);
-    return (
-      <div className={classes.charts}>
-        <div className={classes.chartItem}>
-          <div className={classes.chartContainer}>
-            <div className={classes.chart}>
-              <Chart options={options} series={series} type="donut" width={70} height={70} />
-              <Tooltip title="Pending" placement="top">
-                <Avatar className={classes.iconOverlay} sx={{ bgcolor: 'transparent', width: 24, height: 24 }} style={{ top: 24, left: 23 }}>
-                  <span style={{ color: theme.palette.text?.primary, fontSize: 18 }}><HourglassEmpty style={{ fontSize: 23 }} /></span>
-                </Avatar>
-              </Tooltip>
-            </div>
-          </div>
-          <div className={classes.scoreText}>--%</div>
-          <div className={classes.coverageName}>{t_i18n('Empty coverage')}</div>
-        </div>
-      </div>
-    );
-  }
   return (
-    <div className={classes.charts}>
-      {(coverage_information ?? []).map((coverageResult) => {
-        const { options, series } = genOpts(coverageResult.coverage_score);
-        const warningColor = (theme.palette as { warning?: { main: string } }).warning?.main;
-        let scoreColor;
-        if (coverageResult.coverage_score >= 70) {
-          scoreColor = theme.palette.success.main;
-        } else if (coverageResult.coverage_score >= 40) {
-          scoreColor = warningColor || theme.palette.primary.main;
-        } else {
-          scoreColor = theme.palette.error.main;
-        }
-        return (
-          <div key={coverageResult.coverage_name} className={classes.chartItem}>
-            <div className={classes.chartContainer}>
-              <div className={classes.chart}>
-                <Chart options={options} series={series} type="donut" width={70} height={70} />
-                <Tooltip title={coverageResult.coverage_name} placement="top">
-                  <Avatar className={classes.iconOverlay} sx={{ bgcolor: 'transparent', width: 25, height: 30 }}>
-                    <span style={{ color: theme.palette.text?.primary, fontSize: 18 }}>{getIconFromName(coverageResult.coverage_name, 26)}</span>
-                  </Avatar>
-                </Tooltip>
-              </div>
-            </div>
-            <div className={classes.scoreText} style={{ color: scoreColor }}>
-              {coverageResult.coverage_score}%
-            </div>
-            <div className={classes.coverageName}>
-              {capitalizeFirstLetter(coverageResult.coverage_name)}
-            </div>
-          </div>
-        );
-      })}
-    </div>
+    <Box sx={{ marginBottom: 2 }}>
+      <Label>
+        {t_i18n('Is covered')}
+      </Label>
+      <ItemBoolean
+        status={hasCoverageInformation}
+        label={hasCoverageInformation ? t_i18n('True') : t_i18n('False')}
+      />
+      {hasCoverageInformation && (
+        <>
+          <Label sx={{ marginTop: 2 }}>
+            {t_i18n('Coverage scores')}
+          </Label>
+          <SecurityCoverageScores
+            coverage_information={coverage_information}
+            variant="details"
+          />
+        </>
+      )}
+    </Box>
   );
 };
